move files into sub-packages
