+++ autobahn/src/main/java/xbr/network/crypto/SecretBox.java
@@ -1,4 +1,4 @@
-package xbr.network;
+package xbr.network.crypto;
 
 import org.libsodium.jni.crypto.Random;
 import org.libsodium.jni.crypto.Util;
