+++ autobahn/src/main/java/xbr/network/web3/Pair.java
@@ -10,7 +10,8 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package xbr.network;
+
+package xbr.network.web3;
 
 public class Pair {
     private final Object first;
